fix(sso): fold shared-identity native module into @oxyhq/services; fix cross-app SSO
Silent cross-app "Sign in with Oxy" (shared-keychain handoff) was broken on every
device. Two independent root causes, both verified fixed on-device (Accounts
silently signs in as the identity created in Commons, no QR/prompt):

A. Dependency skew: published @oxyhq/core@9.2.3 pinned @oxyhq/protocol ^0.1.3
   (the version before loadSharedIdentityBridge was added). Consumers resolved a
   stale nested protocol 0.1.3 for core's subtree, so loadSharedIdentityBridge
   was undefined → the JS→native bridge silently resolved null everywhere. Fixed
   by bumping core to 9.2.4 (workspace:^ now emits ^0.1.5).

B. AEADBadTagException: OxyIdentityStore created a fresh EncryptedSharedPreferences
   instance on every read/write. When the writer's instance and a concurrent
   ContentProvider-read instance coexisted, Tink's keyset load raced and the read
   threw AEADBadTagException — turning every cross-app read into "no shared
   identity". Fixed with a single memoized, synchronized instance + commit().

Architecture: the native module no longer lives in a separate @oxyhq/expo-oxy-identity
package (a per-app dep + a version-skew hazard). It now ships inside @oxyhq/services
— which every Oxy app already depends on at a synced version — so it autolinks for
free. @oxyhq/protocol resolves the native module by name (requireOptionalNativeModule
'OxyIdentity'), unchanged. The withSharedIdentityProvider/Reader config plugins moved
to @oxyhq/services/plugins/*; apps reference them there. @oxyhq/expo-oxy-identity is
deleted and dropped from all consumers.
