+++ package.json
@@ -13,7 +13,6 @@
     "packages/core",
     "packages/services",
     "packages/expo-splash",
-    "packages/expo-oxy-identity",
     "packages/api",
     "packages/node",
     "packages/auth",
@@ -36,7 +35,6 @@
     "services:dev": "bun run --filter @oxyhq/services dev",
     "services:test": "bun run --filter @oxyhq/services test",
     "expo-splash:build": "bun run --filter @oxyhq/expo-splash build",
-    "expo-oxy-identity:build": "bun run --filter @oxyhq/expo-oxy-identity build",
     "api:build": "bun run --filter @oxyhq/api build",
     "api:dev": "bun run --filter @oxyhq/api dev",
     "api:start": "bun run --filter @oxyhq/api start",
