+++ packages/accounts/app.config.js
@@ -97,9 +97,10 @@ module.exports = {
       // shared ecosystem cert — enables sign-in-once-use-everywhere.
       './plugins/withSharedUserId',
       // Requests the signature-level READ_IDENTITY permission + provider queries
-      // so this app can READ the shared identity Commons hosts (from
-      // @oxyhq/expo-oxy-identity). Reader-only: it never hosts the provider.
-      './plugins/withSharedIdentityReader',
+      // so this app can READ the shared identity Commons hosts (the native
+      // module now ships inside @oxyhq/services). Reader-only: it never hosts
+      // the provider.
+      '@oxyhq/services/plugins/withSharedIdentityReader',
     ],
     experiments: {
       typedRoutes: true,
