+++ packages/accounts/package.json
@@ -19,7 +19,6 @@
     "@lottiefiles/dotlottie-react": "^0.19.5",
     "@oxyhq/bloom": "^0.29.3",
     "@oxyhq/core": "*",
-    "@oxyhq/expo-oxy-identity": "workspace:*",
     "@oxyhq/expo-splash": "workspace:*",
     "@oxyhq/services": "*",
     "@react-native-async-storage/async-storage": "^2.2.0",
