--- packages/expo-oxy-identity/package.json
@@ -1,79 +1,0 @@
-{
-  "name": "@oxyhq/expo-oxy-identity",
-  "version": "0.1.0",
-  "description": "Cross-app shared Oxy identity bridge for Oxy Expo apps. Commons (the identity vault) writes a single secp256k1 keypair; every same-device Oxy app signed with the shared release key reads it via a signature-protected ContentProvider, enabling silent \"Sign in with Oxy\".",
-  "main": "dist/cjs/index.js",
-  "module": "dist/esm/index.js",
-  "types": "dist/types/index.d.ts",
-  "react-native": "dist/esm/index.js",
-  "source": "src/index.ts",
-  "sideEffects": false,
-  "publishConfig": {
-    "access": "public"
-  },
-  "exports": {
-    ".": {
-      "react-native": "./dist/esm/index.js",
-      "import": {
-        "types": "./dist/types/index.d.ts",
-        "default": "./dist/esm/index.js"
-      },
-      "require": {
-        "types": "./dist/types/index.d.ts",
-        "default": "./dist/cjs/index.js"
-      },
-      "default": "./dist/esm/index.js"
-    },
-    "./package.json": "./package.json"
-  },
-  "files": [
-    "dist",
-    "src",
-    "android",
-    "ios",
-    "expo-module.config.json"
-  ],
-  "keywords": [
-    "oxyhq",
-    "expo",
-    "react-native",
-    "expo-module",
-    "identity",
-    "sso",
-    "content-provider"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/oxyhq/sdk",
-    "directory": "packages/expo-oxy-identity"
-  },
-  "author": "OxyHQ",
-  "license": "MIT",
-  "homepage": "https://oxy.so",
-  "engines": {
-    "node": ">=18.0.0"
-  },
-  "scripts": {
-    "build": "bun run build:cjs && bun run build:esm && bun run build:types",
-    "build:cjs": "tsc -p tsconfig.cjs.json",
-    "build:esm": "tsc -p tsconfig.esm.json && node scripts/fix-esm-imports.mjs",
-    "build:types": "tsc -p tsconfig.types.json",
-    "clean": "rm -rf dist",
-    "typescript": "tsc --noEmit",
-    "test": "jest --passWithNoTests",
-    "prepublishOnly": "bun run clean && bun run build"
-  },
-  "peerDependencies": {
-    "expo": ">=51.0.0",
-    "expo-modules-core": "*",
-    "react": ">=18.0.0",
-    "react-native": ">=0.74.0"
-  },
-  "devDependencies": {
-    "expo": "*",
-    "expo-modules-core": "*",
-    "react": "*",
-    "react-native": "*",
-    "typescript": "^5.9.2"
-  }
-}
