--- packages/expo-oxy-identity/tsconfig.esm.json
@@ -1,14 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "module": "ESNext",
-    "moduleResolution": "bundler",
-    "outDir": "dist/esm",
-    "declaration": false,
-    "noEmit": false,
-    "incremental": true,
-    "tsBuildInfoFile": "dist/esm/.tsbuildinfo"
-  },
-  "include": ["src"],
-  "exclude": ["node_modules", "dist", "**/__tests__"]
-}
