--- packages/expo-oxy-identity/tsconfig.types.json
@@ -1,15 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "module": "ESNext",
-    "moduleResolution": "bundler",
-    "outDir": "dist/types",
-    "declaration": true,
-    "emitDeclarationOnly": true,
-    "noEmit": false,
-    "incremental": true,
-    "tsBuildInfoFile": "dist/types/.tsbuildinfo"
-  },
-  "include": ["src"],
-  "exclude": ["node_modules", "dist", "**/__tests__"]
-}
