+++ packages/services/android/src/main/AndroidManifest.xml
