+++ packages/services/expo-module.config.json
