+++ packages/services/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@oxyhq/services",
-  "version": "19.1.3",
+  "version": "19.2.0",
   "description": "OxyHQ Expo/React Native SDK \u2014 UI components, screens, and native features",
   "main": "lib/commonjs/index.js",
   "module": "lib/module/index.js",
@@ -60,12 +60,18 @@
       },
       "default": "./lib/module/ui/server.js"
     },
+    "./plugins/*": "./plugins/*.js",
+    "./expo-module.config.json": "./expo-module.config.json",
     "./package.json": "./package.json"
   },
   "files": [
     "src",
     "lib",
-    "assets"
+    "assets",
+    "android",
+    "ios",
+    "plugins",
+    "expo-module.config.json"
   ],
   "keywords": [
     "react-native",
@@ -176,6 +182,7 @@
     "@types/react": "*",
     "@types/react-native": "*",
     "expo": ">=56.0.0",
+    "expo-modules-core": "*",
     "expo-document-picker": ">=56.0.0",
     "expo-file-system": ">=56.0.0",
     "expo-font": ">=13.0.0",
@@ -197,6 +204,9 @@
   },
   "peerDependenciesMeta": {
     "nativewind": {
+      "optional": true
+    },
+    "expo-modules-core": {
       "optional": true
     },
     "@expo/vector-icons": {
